# Improve precision and performance for BFloat16 upsampling