--- a/aten/src/ATen/native/UpSampleBicubic2d.cpp
+++ b/aten/src/ATen/native/UpSampleBicubic2d.cpp
@@ -134,10 +134,10 @@ static void upsample_bicubic2d_backward_out_frame(
     }
     return;
   }
-
-  const scalar_t height_scale = area_pixel_compute_scale<scalar_t>(
+  using opmath_t = at::opmath_type<scalar_t>;
+  const opmath_t height_scale = area_pixel_compute_scale<opmath_t>(
       input_height, output_height, align_corners, scales_h);
-  const scalar_t width_scale = area_pixel_compute_scale<scalar_t>(
+  const opmath_t width_scale = area_pixel_compute_scale<opmath_t>(
       input_width, output_width, align_corners, scales_w);
 
   for (const auto output_y : c10::irange(output_height)) {
@@ -145,21 +145,30 @@ static void upsample_bicubic2d_backward_out_frame(
       scalar_t* in = idata;
       scalar_t* out = odata;
 
-      const scalar_t real_x = area_pixel_compute_source_index(width_scale, output_x, align_corners, /*cubic=*/true);
-      int64_t input_x = floorf(real_x);
-      scalar_t t_x = real_x - input_x;
+      const opmath_t real_x = area_pixel_compute_source_index(width_scale, output_x, align_corners, /*cubic=*/true);
+      // when `real_x` becomes larger than the range the floating point
+      // type can accurately represent, the type casting to `int64_t` might exceed
+      // `input_width - 1`. So we guard it with `std::min` below.
+      int64_t input_x = std::min(static_cast<int64_t>(floorf(real_x)), input_width - 1);
+      opmath_t t_x = std::min(
+            std::max(real_x - input_x, static_cast<opmath_t>(0)),
+            static_cast<opmath_t>(1)
+          );
 
-      const scalar_t real_y = area_pixel_compute_source_index(height_scale, output_y, align_corners, /*cubic=*/true);
-      int64_t input_y = floorf(real_y);
-      scalar_t t_y = real_y - input_y;
+      const opmath_t real_y = area_pixel_compute_source_index(height_scale, output_y, align_corners, /*cubic=*/true);
+      int64_t input_y = std::min(static_cast<int64_t>(floorf(real_y)), input_height - 1);
+      opmath_t t_x = std::min(
+            std::max(real_y - input_y, static_cast<opmath_t>(0)),
+            static_cast<opmath_t>(1)
+          );
 
       // NOLINTNEXTLINE(cppcoreguidelines-avoid-c-arrays,modernize-avoid-c-arrays)
-      scalar_t x_coeffs[4];
+      opmath_t x_coeffs[4];
       // NOLINTNEXTLINE(cppcoreguidelines-avoid-c-arrays,modernize-avoid-c-arrays)
-      scalar_t y_coeffs[4];
+      opmath_t y_coeffs[4];
 
-      get_cubic_upsample_coefficients<scalar_t>(x_coeffs, t_x);
-      get_cubic_upsample_coefficients<scalar_t>(y_coeffs, t_y);
+      get_cubic_upsample_coefficients<opmath_t>(x_coeffs, t_x);
+      get_cubic_upsample_coefficients<opmath_t>(y_coeffs, t_y);
 
       for (const auto c C10_UNUSED : c10::irange(channels)) {
         scalar_t out_value = out[output_y * output_width + output_x];