@CaoE 2023-01-18
> Next time, suggest to split the PR (one for precision and one for performance) to ease the code review.

Sorry for the inconvenience. The precision part and performance part are sometimes hard to separate since the codes for improving precision also improve the performance.